# PERF: default_index