Surprisingly big difference here

```
from asv_bench.benchmarks.frame_ctor import *
self3 = FromNDArray()
self3.setup()

%timeit self3.time_frame_from_ndarray()
19.8 µs ± 405 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR
30.9 µs ± 1.38 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- master
```